stock_service/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
pika==1.2.0
beautifulsoup4==4.10.0
